infrastructures/config/test/requirements.txt
fastapi[all]
SQLAlchemy==1.4.48
asyncpg==0.27.0
aiofiles==23.1.0
orjson==3.8.8
httpx==0.23.3
pydantic==1.10.7
passlib==1.7.4
python-dotenv==1.0.0
python-jose==3.3.0

# Testing
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-cov==4.1.0
pytest-env==0.8.1
pytest-mock==3.10.0
pytest-randomly==3.12.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
